remove searching from flowOfFilterQuery
extract `filterByFlagDisplaysProperly` to ViewModelTest
