+++ AnkiDroid/src/test/java/com/ichi2/testutils/MockUtils.kt
@@ -1,0 +1,23 @@
+/*
+ *  Copyright (c) 2024 David Allison <davidallisongithub@gmail.com>
+ *
+ *  This program is free software; you can redistribute it and/or modify it under
+ *  the terms of the GNU General Public License as published by the Free Software
+ *  Foundation; either version 3 of the License, or (at your option) any later
+ *  version.
+ *
+ *  This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ *  PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with
+ *  this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.ichi2.testutils
+
+import org.mockito.Mockito
+import org.mockito.Mockito.RETURNS_DEFAULTS
+
+/** A much which is faster than mockk(relaxed = true) */
+inline fun <reified T> mockIt(): T = Mockito.mock(T::class.java, RETURNS_DEFAULTS)
